requirements.txt
python-dotenv==1.0.0
pandas==2.0.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10